requirements.txt
flask
torch
torchvision
ultralytics
opencv-python-headless
pillow
flask-cors
